test: split NVFP4 Marlin tests into their own file, register for SM120 runner
